web-development/python/flask/job-scheduler/pyproject.toml
[project]
name = "job-scheduler"
version = "0.1.0"
description = "Python Job Scheduler with Flask web dashboard"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "apscheduler>=3.10,<4.0",
    "pyyaml>=6.0",
]
